fix: error not finding id

+++ routes/campgrounds.js
@@ -49,14 +49,28 @@ router.get('/:id', catchAsync (async(req, res) => {
 
 // Edit New campground
 router.get('/:id/edit', isLoggedIn, catchAsync (async(req, res) => {
+    const { id } = req.params
     const campground = await Campground.findById(req.params.id)
+    if (!campground) {
+        req.flash('sccess', 'Canot find that campground')
+        return res.redirect('/campgrounds')
+    }
+    if (!campground.author.equals(req.user._id)) {
+        req.flash('success', 'You dont have permission to do that!')
+        return res.redirect(`/campgrounds/${id}`)
+    } 
     res.render('campgrounds/edit', {campground})
 }))
 
 // PUT Update (Edit) campground
 router.put('/:id', validateCampground, catchAsync (async(req, res) => {
     const { id } = req.params
-    const campground = await Campground.findByIdAndUpdate(id, {...req.body.campground})
+    const campground = await Campground.findById(id)
+    if (!campground.author.equals(req.user._id)) {
+        req.flash('success', 'You dont have permission to do that!')
+        return res.redirect(`/campgrounds/${id}`)
+    }
+    const camp = await Campground.findByIdAndUpdate(id, {...req.body.campground})
     req.flash('success', 'Campground edited successfully')
     res.redirect(`/campgrounds/${campground._id}`)
 }))
@@ -64,6 +78,11 @@ router.put('/:id', validateCampground, catchAsync (async(req, res) => {
 // Delete campground
 router.delete('/:id/', isLoggedIn, catchAsync(async(req, res) => {
     const { id } = req.params
+    const campground = await Campground.findById(id)
+    if (!campground.author.equals(req.user._id)) {
+        req.flash('success', 'You dont have permission to do that!')
+        return res.redirect(`/campground/${id}`)
+    }
     await Campground.findByIdAndDelete(id)
     req.flash('success', 'Campground deleted opps')
     res.redirect('/campgrounds')
